[MODIFY] PR 수정사항 반영 #58
